+++ crates/core/src/parse.rs
@@ -1053,7 +1053,7 @@ static RE_PARSING_PATTERNS: Lazy<ParsingPatterns> = Lazy::new(|| {
     ParsingPatterns {
         extra_inclusion_regex: Regex::new(r#"(?i)^[^/\\]+[/\\]Bindings\.xml$"#).unwrap(),
         initial_inclusion_regex: Regex::new(
-            r#"(?i)^([^/\\]+)[/\\]\1(-mainline|-bcc|-classic)?\.toc$"#,
+            r#"(?i)^([^/\\]+)[/\\]\1([-|_](mainline|bcc|tbc|classic|vanilla))?\.toc$"#,
         )
         .unwrap(),
         file_parsing_regex,