+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and `Removed`.
 
 ### Added
 
+- Classic Era PTR support.
 - A new retry button will appear alongside an error message if one of the API servers
   is down. This new button should make it easier to try fetching the data again
   instead of relaunching the application.