+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and `Removed`.
 
 ## [Unreleased]
 
+### Fixed
+
+- Better error handling if a source goes down.
+
 ## [1.3.1] - 2021-09-12
 
 ### Added