+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and `Removed`.
 
 ## [Unreleased]
 
+
 ### Added
 
 - Option to disable keybindings in Ajour.
@@ -21,6 +22,8 @@ and `Removed`.
 ### Fixed
 
 - Resolved a issue where keybindings would trigger with modifiers.
+- Better handling of the new TOC format with flavor in the filename.
+- Better error handling if a source goes down.
 
 ## [1.3.1] - 2021-09-12